+++ test.js
@@ -1,59 +1,57 @@
-const nodecert = require('.')
+const autoEncryptLocalhost = require('.')
 const test = require('tape')
-const fs = require('fs')
+const fs = require('fs-extra')
 const path = require('path')
 const os = require('os')
 
-// Courtesy: https://gist.github.com/liangzan/807712#gistcomment-337828
-function rmRFSync (dirPath) {
-  try { var files = fs.readdirSync(dirPath) }
-  catch(e) { return }
-  if (files.length > 0)
-    for (var i = 0; i < files.length; i++) {
-      var filePath = dirPath + '/' + files[i]
-      if (fs.statSync(filePath).isFile())
-        fs.unlinkSync(filePath)
-      else
-        rmRFSync(filePath)
-    }
-  fs.rmdirSync(dirPath)
-}
-
 test('certificate creation', t => {
-  t.plan(10)
+  t.plan(12)
 
-  let nodecertDirectory = path.join(os.homedir(), '.small-tech.org', 'nodecert')
-  if (fs.existsSync(nodecertDirectory)) {
-    // Clear the .nodecert directory if it already exists.
-    rmRFSync(nodecertDirectory)
+  const defaultSettingsPath = path.join(os.homedir(), '.small-tech.org', 'auto-encrypt-localhost')
+
+  const keyFilePath = path.join(defaultSettingsPath, 'localhost-key.pem')
+  const certFilePath = path.join(defaultSettingsPath, 'localhost.pem')
+
+  // Clear the settings path if it already exists.
+  if (fs.existsSync(defaultSettingsPath)) {
+    fs.removeSync(defaultSettingsPath)
   }
 
-  // Run nodecert.
-  nodecert()
+  // Run Auto Encrypt Localhost.
+  const tlsOptions = autoEncryptLocalhost()
 
-  t.ok(fs.existsSync(path.join(nodecertDirectory)), 'Main nodecert directory exists')
-  t.ok(fs.existsSync(path.join(nodecertDirectory, 'rootCA.pem')), 'Local certificate authority exists')
-  t.ok(fs.existsSync(path.join(nodecertDirectory, 'rootCA-key.pem')), 'Local certificate authority private key exists')
-  t.ok(fs.existsSync(path.join(nodecertDirectory, 'localhost.pem')), 'Local certificate exists')
-  t.ok(fs.existsSync(path.join(nodecertDirectory, 'localhost-key.pem')), 'Local certificate private key exists')
+  t.ok(fs.existsSync(path.join(defaultSettingsPath)), 'Main settings path exists.')
+  t.ok(fs.existsSync(path.join(defaultSettingsPath, 'rootCA.pem')), 'Local certificate authority exists.')
+  t.ok(fs.existsSync(path.join(defaultSettingsPath, 'rootCA-key.pem')), 'Local certificate authority private key exists.')
+  t.ok(fs.existsSync(certFilePath), 'Local certificate exists.')
+  t.ok(fs.existsSync(keyFilePath), 'Local certificate private key exists.')
+
+  // Ensure that the certificate and private key in the returned tls options object matches
+  // what exists on the file system.
+  const key = fs.readFileSync(keyFilePath, 'utf-8')
+  const cert = fs.readFileSync(certFilePath, 'utf-8')
+
+  t.strictEquals(tlsOptions.key, key, 'Private key returned in tls options object matches key from disk.')
+  t.strictEquals(tlsOptions.cert, cert, 'Certificate returned in tls options object matches key from disk.')
 
   //
-  // Custom nodecert directory.
+  // Custom settings path.
   //
 
-  nodecertDirectory = path.join(os.homedir(), '.small-tech.org', 'nodecert-custom-directory-test', 'second-level-directory')
-  if (fs.existsSync(nodecertDirectory)) {
-    // Clear the .nodecert directory if it already exists.
-    rmRFSync(nodecertDirectory)
+  const customSettingsPath = path.join(os.homedir(), '.small-tech.org', 'auto-encrypt-localhost-custom-directory-test', 'second-level-directory')
+
+  // Clear the custom settings path if it already exists.
+  if (fs.existsSync(customSettingsPath)) {
+    fs.removeSync(customSettingsPath)
   }
 
-  nodecert(nodecertDirectory)
+  autoEncryptLocalhost({ settingsPath: customSettingsPath })
 
-  t.ok(fs.existsSync(path.join(nodecertDirectory)), '(Custom path) Main nodecert directory exists')
-  t.ok(fs.existsSync(path.join(nodecertDirectory, 'rootCA.pem')), '(Custom path) Local certificate authority exists')
-  t.ok(fs.existsSync(path.join(nodecertDirectory, 'rootCA-key.pem')), '(Custom path) Local certificate authority private key exists')
-  t.ok(fs.existsSync(path.join(nodecertDirectory, 'localhost.pem')), '(Custom path) Local certificate exists')
-  t.ok(fs.existsSync(path.join(nodecertDirectory, 'localhost-key.pem')), '(Custom path) Local certificate private key exists')
+  t.ok(fs.existsSync(path.join(customSettingsPath)), '(Custom settings path) Main directory exists.')
+  t.ok(fs.existsSync(path.join(customSettingsPath, 'rootCA.pem')), '(Custom settings path) Local certificate authority exists.')
+  t.ok(fs.existsSync(path.join(customSettingsPath, 'rootCA-key.pem')), '(Custom settings path) Local certificate authority private key exists.')
+  t.ok(fs.existsSync(path.join(customSettingsPath, 'localhost.pem')), '(Custom settings path) Local certificate exists.')
+  t.ok(fs.existsSync(path.join(customSettingsPath, 'localhost-key.pem')), '(Custom settings path) Local certificate private key exists.')
 
   t.end()
 })